rtl: Added sdram_ctrl.sv and sdram test
More tests are required the sdram controller. The sdram simulator is
very limited on memory since it's implemented in Verilog and large RAMs
can lower Verilator's perf a lot. In the future moving at least the
memory implementation to C++ will be required.

Also improved the Verilator test Makefile. A lot of future work is still
required though.